+++ src/components/SharedComponents/CardList.jsx
@@ -15,7 +15,7 @@ function CardList() {
   const { data, isLoading, error } = useFetch('api/sample/folder');
 
   if (isLoading) {
-    return <div>Loading...</div>;
+    return <div></div>;
   }
 
   if (error) {
