+++ src/components/SharedComponents/Content.jsx
@@ -1,4 +1,4 @@
-import Search from './Search';
+import Search from '../Search';
 import CardList from './CardList';
 
 function Content() {
