+++ src/routes/Shared.jsx
@@ -1,7 +1,7 @@
-import Nav from '../components/SharedComponents/Nav.jsx';
+import Nav from '../components/Nav.jsx';
 import Header from '../components/SharedComponents/Header.jsx';
 import Content from '../components/SharedComponents/Content.jsx';
-import Footer from '../components/SharedComponents/Footer.jsx';
+import Footer from '../components/Footer.jsx';
 
 function Shared() {
   return (
